Complete reorganize, made both projects live happily side-by-side
